+++ processor/src/fast/io_ops.rs
@@ -1,5 +1,5 @@
 use super::{DOUBLE_WORD_SIZE, ExecutionError, FastProcessor, Felt, WORD_SIZE_FELT};
-use crate::{ErrorContext, Host, ProcessState};
+use crate::{ErrorContext, Host};
 
 impl FastProcessor {
     /// Analogous to `Process::op_push`.
@@ -12,7 +12,8 @@ impl FastProcessor {
     pub fn op_advpop(&mut self, op_idx: usize, host: &mut impl Host) -> Result<(), ExecutionError> {
         let value = host
             .advice_provider_mut()
-            .pop_stack(ProcessState::new_fast(self, op_idx), &ErrorContext::default())?;
+            .pop_stack()
+            .map_err(|err| err.into_exec_err_at_clk(self.clk + op_idx, &ErrorContext::default()))?;
         self.increment_stack_size();
         self.stack_write(0, value);
         Ok(())
@@ -26,7 +27,8 @@ impl FastProcessor {
     ) -> Result<(), ExecutionError> {
         let word = host
             .advice_provider_mut()
-            .pop_stack_word(ProcessState::new_fast(self, op_idx), &ErrorContext::default())?;
+            .pop_stack_word()
+            .map_err(|err| err.into_exec_err_at_clk(self.clk + op_idx, &ErrorContext::default()))?;
         self.stack_write_word(0, &word);
 
         Ok(())
@@ -111,7 +113,8 @@ impl FastProcessor {
         // pop two words from the advice stack
         let words = host
             .advice_provider_mut()
-            .pop_stack_dword(ProcessState::new_fast(self, op_idx), &ErrorContext::default())?;
+            .pop_stack_dword()
+            .map_err(|err| err.into_exec_err_at_clk(self.clk + op_idx, &ErrorContext::default()))?;
 
         // write the words to memory
         self.memory.write_word(self.ctx, addr_first_word, self.clk + op_idx, words[0])?;