+++ processor/src/operations/io_ops.rs
@@ -222,7 +222,7 @@ impl Process {
     pub(super) fn op_pipe(
         &mut self,
         host: &mut impl Host,
-        error_ctx: &ErrorContext<'_, impl MastNodeExt>,
+        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
     ) -> Result<(), ExecutionError> {
         const MEM_ADDR_STACK_IDX: usize = 12;
 
@@ -233,16 +233,19 @@ impl Process {
         let addr_second_word = addr_first_word + Felt::from(WORD_SIZE as u32);
 
         // pop two words from the advice stack
-        let words = host.advice_provider_mut().pop_stack_dword(self.into(), error_ctx)?;
+        let words = host
+            .advice_provider_mut()
+            .pop_stack_dword()
+            .map_err(|err| err.into_exec_err_at_clk(clk, err_ctx))?;
 
         // write the words memory
         self.chiplets
             .memory
-            .write_word(ctx, addr_first_word, clk, words[0], error_ctx)
+            .write_word(ctx, addr_first_word, clk, words[0], err_ctx)
             .map_err(ExecutionError::MemoryError)?;
         self.chiplets
             .memory
-            .write_word(ctx, addr_second_word, clk, words[1], error_ctx)
+            .write_word(ctx, addr_second_word, clk, words[1], err_ctx)
             .map_err(ExecutionError::MemoryError)?;
 
         // replace the elements on the stack with the word elements (in stack order)
@@ -278,7 +281,10 @@ impl Process {
         host: &mut impl Host,
         err_ctx: &ErrorContext<'_, impl MastNodeExt>,
     ) -> Result<(), ExecutionError> {
-        let value = host.advice_provider_mut().pop_stack(self.into(), err_ctx)?;
+        let value = host
+            .advice_provider_mut()
+            .pop_stack()
+            .map_err(|err| err.into_exec_err_at_clk(self.system.clk(), err_ctx))?;
         self.stack.set(0, value);
         self.stack.shift_right(0);
         Ok(())
@@ -294,7 +300,10 @@ impl Process {
         host: &mut impl Host,
         err_ctx: &ErrorContext<'_, impl MastNodeExt>,
     ) -> Result<(), ExecutionError> {
-        let word = host.advice_provider_mut().pop_stack_word(self.into(), err_ctx)?;
+        let word = host
+            .advice_provider_mut()
+            .pop_stack_word()
+            .map_err(|err| err.into_exec_err_at_clk(self.system.clk(), err_ctx))?;
 
         self.stack.set(0, word[3]);
         self.stack.set(1, word[2]);
@@ -607,7 +616,6 @@ mod tests {
 
     #[test]
     fn op_pipe() {
-        let err_ctx = ErrorContext::default();
         let mut host = DefaultHost::default();
         let mut process = Process::new_dummy_with_decoder_helpers_and_empty_stack();
         let program = &MastForest::default();
@@ -619,9 +627,7 @@ mod tests {
         let word2_felts: [Felt; WORD_SIZE] = word2.to_elements().try_into().unwrap();
         for element in word2_felts.iter().rev().chain(word1_felts.iter().rev()).copied() {
             // reverse the word order, since elements are pushed onto the advice stack.
-            host.advice_provider_mut()
-                .push_stack(AdviceSource::Value(element), &err_ctx)
-                .unwrap();
+            host.advice_provider_mut().push_stack(AdviceSource::Value(element)).unwrap();
         }
 
         // arrange the stack such that: