+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - [BREAKING] Updated dependencies Winterfell to v0.13 and Crypto to v0.15 (#1896).
 - Fixed instructions with errors print without quotes (#1882).
 - [BREAKING] Convert `AdviceProvider` into a struct ([#1904](https://github.com/0xMiden/miden-vm/issues/1904), [#1905](https://github.com/0xMiden/miden-vm/issues/1905))
+- [BREAKING] `AdviceProvider` should return its own error type ([#1907](https://github.com/0xMiden/miden-vm/issues/1907)
 
 #### Enhancements
 
@@ -28,7 +29,6 @@
 
 - Modules can now be provided in any order to the `Assembler`, see #1669 (#1881)
 - Addressed bug which caused references to re-exported procedures whose definition internally referred to an aliased module import, to produce an "undefined module" error, see #1451 (#1892)
-
 
 ## 0.15.0 (2025-06-06)